requirements.txt
bleach==4.1.0
bokeh==2.4.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorcet==2.0.6
cycler==0.10.0
Flask==2.0.1
Flask-Cors==3.0.10
holoviews==1.14.6
hvplot==0.7.3
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
packaging==21.0
pandas==1.3.3
panel==0.12.3
param==1.11.1
Pillow==8.3.2
pyct==0.4.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pyviz-comms==2.1.0
PyYAML==5.4.1
requests==2.26.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.1
